+++ ui/src/App.js
@@ -1,26 +1,13 @@
 import React, { Component } from 'react';
-import logo from './logo.svg';
 import './App.css';
-import {app} from "./config/fire";
 import firebase from "firebase";
 import axios from "axios";
 import { BrowserRouter as Router, Route } from "react-router-dom";
-import { AuthProvider } from "./Auth";
-import Login from "./Login";
-import Home from "./Home";
-import PrivateRoute from "./PrivateRoute";
+import { AuthProvider } from "./components/Auth/Auth";
+import Login from "./components/Login/Login";
+import Home from "./components/Home/Home";
+import PrivateRoute from "./components/PrivateRoute";
 
-const auth = app.auth();
-let provider = new firebase.auth.GoogleAuthProvider();
-
-const signInWithGoogle = () => {
-  auth.signInWithPopup(provider).then(function(res) {
-
-    let user = res.user;
-    console.log(user);
-    console.log(user.uid);
-  })
-};
 
 class App extends Component {
   constructor (props){
@@ -38,11 +25,13 @@ class App extends Component {
           const heartrate = res.data;
                 console.log(heartrate)
          })
+       } else {
+         this.setState({user: null})
        }
     })
  }
-  render () {
 
+  render () {
   return (
     <AuthProvider>
       <Router>
