--- ui/src/Home.js
@@ -1,13 +1,0 @@
-import React from "react";
-import {app} from "./config/fire";
-
-const Home = () => {
-  return (
-    <>
-      <h1>Home</h1>
-      <button onClick={() => app.auth().signOut()}>Sign out</button>
-    </>
-  );
-};
-
-export default Home;
